rename submission script, add section on submitting libe as balsam app to readme
